Progress

+++ philo/src/time.c
@@ -6,13 +6,13 @@
 /*   By: csilva-m <csilva-m@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/30 17:46:49 by csilva-m          #+#    #+#             */
-/*   Updated: 2024/07/30 18:07:26 by csilva-m         ###   ########.fr       */
+/*   Updated: 2024/08/13 17:12:41 by csilva-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-long long get_time(void)
+size_t get_time(void)
 {
 	struct timeval	time;
 
@@ -20,7 +20,7 @@ long long get_time(void)
 	return ((time.tv_sec * 1000) + (time.tv_usec / 1000));
 }
 
-long long get_ms()
+size_t get_ms()
 {
 	return(get_time() - get_core()->day);
 }
